+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The changelog is available [on GitHub][2].
 =====
 * Fix builds with newer primitive.
 * Add `extend` function
+* Add `unsafeLatest` function
 
 0.0.0
 =====